requirements.txt
fastapi
uvicorn[standard]
python-multipart
langchain
langchain-google-genai
langchain-community
chromadb
pypdf
streamlit
requests
python-dotenv
